requirements.txt
streamlit
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
